requirements.txt
google-genai>=1.68.0
openai>=2.29.0
smolagents>=1.24.0
litellm>=1.82.0,<=1.82.6
requests>=2.32.0
httpx>=0.28.0
